Make Comdb2 server build without SSL/TLS support.
