+++ CMakeLists.txt
@@ -17,6 +17,13 @@ if(COMDB2_BUILD_STATIC)
   SET(CMAKE_FIND_LIBRARY_SUFFIXES .a ${CMAKE_FIND_LIBRARY_SUFFIXES})
 endif()
 
+option(WITH_SSL "Turn OFF to compile without SSL" ON)
+if(WITH_SSL)
+  add_definitions(-DWITH_SSL=1)
+else()
+  add_definitions(-DWITH_SSL=0)
+endif()
+
 if(${CMAKE_SYSTEM_NAME} STREQUAL Darwin)
   # Homebrew will install here by default
   list(APPEND CMAKE_PREFIX_PATH
@@ -29,6 +36,7 @@ set(CMAKE_MODULE_PATH ${CMAKE_MODULE_PATH} ${PROJECT_SOURCE_DIR}/cmake)
 set(CURSES_NEED_NCURSES TRUE)
 find_package(Curses REQUIRED)
 find_package(LZ4 REQUIRED)
+# Comdb2 uses libcrpyto. Hence OpenSSL is required even if WITH_SSL is off.
 find_package(OpenSSL REQUIRED)
 find_package(Protobuf_C REQUIRED)
 find_package(Readline REQUIRED)
