+++ bb/sbuf2.c
@@ -27,6 +27,9 @@
 #include <strings.h>
 #include <sys/uio.h>
 #include <unistd.h>
+#include <netdb.h>
+#include <netinet/in.h>
+#include <arpa/inet.h>
 
 #if SBUF2_SERVER
 #  ifndef SBUF2_DFL_SIZE
@@ -810,6 +813,178 @@ void SBUF2_FUNC(sbuf2setuserptr)(SBUF2 *sb, void *userptr)
 void *SBUF2_FUNC(sbuf2getuserptr)(SBUF2 *sb)
 {
     return sb->userptr;
+}
+
+#if SBUF2_SERVER
+#include <lockmacro.h> /* LOCK & UNLOCK */
+#include <plhash.h>    /* hash_t */
+#include <logmsg.h>    /* logmsg */
+#define logi(...) logmsg(LOGMSG_INFO, ##__VA_ARGS__)
+#define loge(...) logmsg(LOGMSG_ERROR, ##__VA_ARGS__)
+static pthread_mutex_t peer_lk = PTHREAD_MUTEX_INITIALIZER;
+static hash_t *peer_hash = NULL;
+struct peer_info {
+    /* Key (must come first in struct) */
+    struct in_addr addr;
+    short family;
+    /* (don't forget to bzero the key area as the member fields may not
+     * be aligned) */
+
+    /* Data */
+    char *host;
+};
+#else
+/* Keep quiet in client mode. */
+#define logi(...)
+#define loge(...)
+/* Don't cache host info in client mode. */
+#ifdef LOCK
+#undef LOCK
+#endif /* LOCK */
+#define LOCK(arg)
+#ifdef UNLOCK
+#undef UNLOCK
+#endif /* UNLOCK */
+#define UNLOCK(arg)
+#endif /* SBUF2_SERVER */
+
+char *SBUF2_FUNC(get_origin_mach_by_buf)(SBUF2 *sb)
+{
+    int fd;
+    char *funcname;
+    struct sockaddr_in peeraddr;
+    socklen_t len = sizeof(peeraddr);
+    char *host;
+#if SBUF2_SERVER
+    struct peer_info *info;
+    struct peer_info key;
+#else
+    void *info = NULL;
+#endif
+
+    if (sb == NULL)
+        return "???";
+
+    fd = sb->fd;
+
+    if (fd == -1)
+        return "???";
+
+    bzero(&peeraddr, sizeof(peeraddr));
+    if (getpeername(fd, (struct sockaddr *)&peeraddr, &len) < 0) {
+        loge("%s:getpeername failed fd %d: %d %s\n", __func__, fd, errno,
+             strerror(errno));
+        return "???";
+    }
+
+#if SBUF2_SERVER
+    bzero(&key, offsetof(struct peer_info, host));
+    memcpy(&key.addr, &peeraddr.sin_addr, sizeof(key.addr));
+    key.family = peeraddr.sin_family;
+#endif
+
+    LOCK(&peer_lk)
+    {
+#if SBUF2_SERVER
+        if (!peer_hash) {
+            peer_hash = hash_init(offsetof(struct peer_info, host));
+            if (!peer_hash) {
+                loge("%s:hash_init failed\n", __func__);
+                errUNLOCK(&peer_lk);
+                return "???";
+            }
+        }
+
+        info = hash_find(peer_hash, &key);
+#endif
+        if (!info) {
+            /* Do a slow lookup of this internet address in the host database
+             * to get a hostname, and then search the bigsnd node list to
+             * map this to a node number. */
+            struct hostent *hp = NULL, rslt;
+            char hnm[256] = {0};
+            char *h_name = NULL;
+            int node, rc;
+            int error_num = 0;
+            int goodrc = 0;
+
+#ifdef _LINUX_SOURCE
+            funcname = "getnameinfo";
+            rc = getnameinfo((struct sockaddr *)&peeraddr, sizeof(peeraddr),
+                             hnm, sizeof(hnm), NULL, 0, 0);
+
+            if (0 == rc) {
+                goodrc = 1;
+                h_name = hnm;
+            } else {
+                error_num = errno;
+            }
+#else
+            funcname = "getipnodebyaddr";
+            hp = getipnodebyaddr(&peeraddr.sin_addr, sizeof(peeraddr.sin_addr),
+                                 peeraddr.sin_family, &error_num);
+            if (hp) {
+                goodrc = 1;
+                h_name = hp->h_name;
+            }
+#endif
+
+            if (0 == goodrc) {
+                char addrstr[64] = "";
+                inet_ntop(peeraddr.sin_family, &peeraddr.sin_addr, addrstr,
+                          sizeof(addrstr));
+                loge("%s:%s failed fd %d (%s): error_num %d", __func__,
+                     funcname, fd, addrstr, error_num);
+                switch (error_num) {
+                case HOST_NOT_FOUND:
+                    loge(" HOST_NOT_FOUND\n");
+                    break;
+                case NO_DATA:
+                    loge(" NO_DATA\n");
+                    break;
+                case NO_RECOVERY:
+                    loge(" NO_RECOVERY\n");
+                    break;
+                case TRY_AGAIN:
+                    loge(" TRY_AGAIN\n");
+                    break;
+                default:
+                    loge(" ???\n");
+                    break;
+                }
+                host = strdup(addrstr);
+            } else {
+                host = strdup(h_name);
+            }
+
+#if SBUF2_SERVER
+            info = calloc(1, sizeof(struct peer_info));
+            if (!info) {
+                errUNLOCK(&peer_lk);
+                loge("%s: out of memory\n", __func__);
+                free(host);
+                host = NULL;
+                return "???";
+            }
+
+            memcpy(info, &key, sizeof(key));
+            info->host = host;
+            if (hash_add(peer_hash, info) != 0) {
+                errUNLOCK(&peer_lk);
+                loge("%s: hash_add failed\n", __func__);
+                free(info);
+                return host;
+            }
+#endif
+        }
+    }
+    UNLOCK(&peer_lk);
+
+#if SBUF2_SERVER
+    return (info != NULL) ? info->host : "???";
+#else
+    return (host != NULL) ? host : "???";
+#endif
 }
 
 #if WITH_SSL
