+++ cdb2api/cdb2api.c
@@ -3388,7 +3388,6 @@ read_record:
                 (void *)pthread_self(), host, __LINE__, rc, type);
     }
 
-#if WITH_SSL
     if (type == RESPONSE_HEADER__SQL_RESPONSE_SSL) {
 #if WITH_SSL
         hndl->s_sslmode = PEER_SSL_REQUIRE;
@@ -3397,12 +3396,13 @@ read_record:
 
         /* Decrement retry counter: It is not a real retry. */
         --retries_done;
+        GOTO_RETRY_QUERIES();
 #else
+        sprintf(hndl->errstr, "%s: The database requires SSL connections.",
+                __func__);
         PRINT_RETURN(-1);
 #endif
-        GOTO_RETRY_QUERIES();
     }
-#endif
 
     /* Dbinfo .. go to new node */
     if (type == RESPONSE_HEADER__DBINFO_RESPONSE) {
@@ -4665,7 +4665,7 @@ done:
 }
 
 #if WITH_SSL
-#  include <ssl_support.c>
+#include <ssl_support.h>
 static int set_up_ssl_params(cdb2_hndl_tp *hndl)
 {
     /* In case that the application connects to multiple databases
