+++ db/fdb_comm.c
@@ -1213,11 +1213,13 @@ static int fdb_msg_read_message(SBUF2 *sb, fdb_msg_t *msg,
     int isuuid = 0;
     int recv_dk = 0;
 
+#if WITH_SSL
     if (gbl_client_ssl_mode >= SSL_REQUIRE && !gbl_ssl_allow_remsql) {
         logmsg(LOGMSG_ERROR,
                "Remote SQL is forbidden because client SSL is required.");
         return -1;
     }
+#endif
 
     /* clean previous message */
     fdb_msg_clean_message(msg);
