+++ tools/cdb2sql/CMakeLists.txt
@@ -10,13 +10,16 @@ include_directories(
   ${OPENSSL_INCLUDE_DIR}
   ${READLINE_INCLUDE_DIR}
 )
-target_link_libraries(cdb2sql
+set(libs
   cdb2api
-  ${OPENSSL_LIBRARIES}
   ${PROTOBUF_C_LIBRARY}
   ${READLINE_LIBRARY}
   ${CURSES_LIBRARIES}
   ${ZLIB_LIBRARIES}
   ${CMAKE_DL_LIBS}
 )
+if(WITH_SSL)
+  list(APPEND libs ${OPENSSL_LIBRARIES})
+endif()
+target_link_libraries(cdb2sql ${libs})
 install(TARGETS cdb2sql RUNTIME DESTINATION bin)
