+++ tools/pmux/CMakeLists.txt
@@ -12,15 +12,18 @@ include_directories(
   ${PROJECT_SOURCE_DIR}/sockpool
   ${SQLITE3_INCLUDE_DIR}
 )
-target_link_libraries(pmux
+set(libs
   cdb2api
   sockpool
-  ${OPENSSL_LIBRARIES}
   ${PROTOBUF_C_LIBRARY}
   ${SQLITE3_LIBRARY}
   ${ZLIB_LIBRARIES}
   ${CMAKE_DL_LIBS}
 )
+if(WITH_SSL)
+  list(APPEND libs ${OPENSSL_LIBRARIES})
+endif()
+target_link_libraries(pmux ${libs})
 if(COMDB2_BUILD_STATIC)
   set(CMAKE_EXE_LINKER_FLAGS "${CMAKE_EXE_LINKER_FLAGS} -static-libgcc -static-libstdc++")
 endif()
